FSE_Embedded/Trabalho 2 - Controle de AirFryer/requirements.txt
RPi.GPIO
smbus
bmp280
serial
